Move translate-toolkit specific code to TTKitFormat class
This way we can now start creating formats independent of
translate-toolkit.
