+++ weblate/formats/base.py
@@ -21,8 +21,6 @@
 
 from __future__ import unicode_literals
 
-import importlib
-import inspect
 import os
 import re
 import sys
@@ -30,13 +28,6 @@ import tempfile
 
 from django.utils.functional import cached_property
 
-import six
-
-from translate.storage.base import TranslationStore
-from translate.storage.lisa import LISAfile
-from translate.storage.properties import propfile
-
-from weblate.utils.errors import report_error
 from weblate.utils.hash import calculate_hash
 
 FLAGS_RE = re.compile(r'\b[-\w:]+\b')
@@ -179,7 +170,6 @@ class TranslationFormat(object):
     """Generic object defining file format loader."""
     name = ''
     format_id = ''
-    loader = ('', '')
     monolingual = None
     check_flags = ()
     unit_class = TranslationUnit
@@ -191,66 +181,19 @@ class TranslationFormat(object):
     def get_identifier(cls):
         return cls.format_id
 
-    @staticmethod
-    def serialize(store):
-        """Serialize given ttkit store"""
-        return bytes(store)
-
     @classmethod
     def parse(cls, storefile, template_store=None, language_code=None):
-        """Parse store and returns TranslationFormat instance."""
+        """Parse store and returns TranslationFormat instance.
+
+        This wrapper is needed for AutoFormat to be able to return
+        instance of different class."""
         return cls(storefile, template_store, language_code)
-
-    @classmethod
-    def fixup(cls, store):
-        """Perform optional fixups on store."""
-        return store
-
-    @classmethod
-    def load(cls, storefile):
-        """Load file using defined loader."""
-        # Add missing mode attribute to Django file wrapper
-        if (not isinstance(storefile, six.string_types) and
-                not hasattr(storefile, 'mode')):
-            storefile.mode = 'r'
-
-        return cls.parse_store(storefile)
-
-    @classmethod
-    def get_class(cls):
-        """Return class for handling this module."""
-        # Direct class
-        if inspect.isclass(cls.loader):
-            return cls.loader
-        # Tuple style loader, import from translate toolkit
-        module_name, class_name = cls.loader
-        if '.' not in module_name:
-            module_name = 'translate.storage.{0}'.format(module_name)
-        module = importlib.import_module(module_name)
-
-        # Get the class
-        return getattr(module, class_name)
-
-    @classmethod
-    def parse_store(cls, storefile):
-        """Parse the store."""
-        storeclass = cls.get_class()
-
-        # Parse file
-        store = storeclass.parsefile(storefile)
-
-        # Apply possible fixups and return
-        return cls.fixup(store)
 
     def __init__(self, storefile, template_store=None, language_code=None):
         """Create file format object, wrapping up translate-toolkit's store."""
         self.storefile = storefile
         # Load store
-        if isinstance(storefile, TranslationStore):
-            # Used by XLSX writer
-            self.store = storefile
-        else:
-            self.store = self.load(storefile)
+        self.store = self.load(storefile)
         # Check store validity
         if not self.is_valid(self.store):
             raise ValueError(
@@ -258,10 +201,10 @@ class TranslationFormat(object):
             )
         # Remember template
         self.template_store = template_store
-        # Set language (needed for some which do not include this)
-        if (language_code is not None and
-                self.store.gettargetlanguage() is None):
-            self.store.settargetlanguage(language_code)
+
+    @classmethod
+    def load(cls, storefile):
+        raise NotImplementedError()
 
     def get_plural(self, language):
         """Return matching plural object."""
@@ -329,11 +272,7 @@ class TranslationFormat(object):
 
     def add_unit(self, ttkit_unit):
         """Add new unit to underlaying store."""
-        if isinstance(self.store, LISAfile):
-            # LISA based stores need to know this
-            self.store.addunit(ttkit_unit, new=True)
-        else:
-            self.store.addunit(ttkit_unit)
+        raise NotImplementedError()
 
     def update_header(self, **kwargs):
         """Update store header if available."""
@@ -341,7 +280,7 @@ class TranslationFormat(object):
 
     def save_content(self, handle):
         """Stores content to file."""
-        self.store.serialize(handle)
+        raise NotImplementedError()
 
     def save(self):
         """Save underlaying store to disk."""
@@ -374,44 +313,22 @@ class TranslationFormat(object):
     @property
     def mimetype(self):
         """Return most common mime type for format."""
-        if self.store.Mimetypes is None:
-            # Properties files do not expose mimetype
-            return 'text/plain'
-        return self.store.Mimetypes[0]
+        return 'text/plain'
 
     @property
     def extension(self):
         """Return most common file extension for format."""
-        if self.store.Extensions is None:
-            return 'txt'
-        return self.store.Extensions[0]
+        return 'txt'
 
     @classmethod
     def is_valid(cls, store):
-        """Check whether store seems to be valid.
-
-        In some cases ttkit happily "parses" the file, even though it
-        really did not do so (eg. Gettext parser on random text file).
-        """
-        if store is None:
-            return False
-
-        if cls.monolingual is False and not store.units:
-            return False
-
+        """Check whether store seems to be valid."""
         return True
 
     @classmethod
     def is_valid_base_for_new(cls, base, monolingual):
         """Check whether base is valid."""
-        if not base:
-            return monolingual and cls.new_translation is not None
-        try:
-            cls.parse_store(base)
-            return True
-        except Exception as error:
-            report_error(error)
-            return False
+        raise NotImplementedError()
 
     @staticmethod
     def get_language_code(code):
@@ -439,16 +356,7 @@ class TranslationFormat(object):
     @classmethod
     def create_new_file(cls, filename, language, base):
         """Handle creation of new translation file."""
-        if base:
-            # Parse file
-            store = cls.parse_store(base)
-            cls.untranslate_store(store, language)
-            store.savefile(filename)
-        elif cls.new_translation is None:
-            raise ValueError('Not supported')
-        else:
-            with open(filename, 'w') as output:
-                output.write(cls.new_translation)
+        raise NotImplementedError()
 
     def iterate_merge(self, fuzzy):
         """Iterate over units for merging.
@@ -478,28 +386,11 @@ class TranslationFormat(object):
 
     @classmethod
     def untranslate_store(cls, store, language, fuzzy=False):
-        """Remove translations from ttkit store"""
-        store.settargetlanguage(language.code)
-        plural = language.plural
-
-        for unit in store.units:
-            if unit.istranslatable():
-                if hasattr(unit, 'markapproved'):
-                    # Xliff only
-                    unit.markapproved(False)
-                else:
-                    unit.markfuzzy(fuzzy)
-                if unit.hasplural():
-                    unit.target = [''] * plural.number
-                else:
-                    unit.target = ''
+        """Remove translations from a store"""
+        raise NotImplementedError()
 
     def create_unit(self, key, source):
-        unit = self.store.UnitClass(source)
-        unit.setid(key)
-        unit.source = key
-        unit.target = source
-        return unit
+        raise NotImplementedError()
 
     def new_unit(self, key, source):
         """Add new unit to monolingual store."""
